Export

Export data out of R and write to disk.

export(object, ...)

# S4 method for FGSEAList
export(object, name = NULL, dir = ".", geneSetResults = FALSE)

Arguments

object
Object.

name
character(1). Name to use on disk. If NULL, will use the name of the object instead.

dir
character(1). Directory path.

geneSetResults
logical(1). Export expression CSV files per gene set.

...
Additional arguments.

Value

Invisibily return file paths, either as list or character.

Note

Updated 2020-09-23.

On-disk structure

Example: fgsea/mutant_vs_control/c1.csv

S4 object is currently structured by:

Gene set (c1-c8, h).

Contrast.

The object was structured in this manner to flow with the R Markdown template. However, when writing to disk, I think it makes more sense to organize by:

Contrast

Gene set.

Examples

data(fgsea)

## FGSEAList ====
export(
    object = fgsea,
    dir = "example",
    geneSetResults = FALSE
)

#> → Exporting to '/Users/mike/git/monorepo/r-packages/acidgsea/docs/reference/example/fgsea'.

#> → Exporting h.csv at /Users/mike/git/monorepo/r-packages/acidgsea/docs/reference/example/fgsea/condition_B_vs_A using vroom::`vroom_write()`.

#> → Exporting h.csv at /Users/mike/git/monorepo/r-packages/acidgsea/docs/reference/example/fgsea/treatment_D_vs_C using vroom::`vroom_write()`.

sort(list.files(file.path("example", "fgsea")))

#> [1] "condition_B_vs_A" "treatment_D_vs_C"

unlink("example", recursive = TRUE)