Top tables

Top tables of significantly enriched pathways.

topTables(object, ...)

# S4 method for FGSEAList
topTables(object, collection, n = 10L, headerLevel = 3L)

Arguments

object
Object.

collection
character(1) or integer(1). Gene set collection name or position, corresponding to values defined in names(). Typically refers to h (hallmark), c1-c7 collections from MSigDb. Can obtain using collectionNames() on FGSEAList object.

n
integer(1). Number of significant processes to include. Bidirectional; will include n up- and down-regulated processes.

headerLevel
integer(1). Markdown header level.

...
Additional arguments.

Value

Markdown output.

Details

Supports looping across multiple DEG results, and adds a Markdown header for each contrast.

Note

Updated 2020-08-05.

Examples

data(fgsea)

## FGSEAList ====
alphaThreshold(fgsea) <- 0.9
topTables(fgsea, collection = "h")

#> 
#> 
#> ### condition_B_vs_A
#> 
#> 
#> 
#> #### Upregulated
#> 
#> 
#> 
#> |pathway                            |  padj|   NES| size|
#> |:----------------------------------|-----:|-----:|----:|
#> |HALLMARK_PANCREAS_BETA_CELLS       | 0.672| 1.397|   15|
#> |HALLMARK_NOTCH_SIGNALING           | 0.672| 1.371|   17|
#> |HALLMARK_P53_PATHWAY               | 0.672| 1.336|   63|
#> |HALLMARK_MITOTIC_SPINDLE           | 0.672| 1.239|   77|
#> |HALLMARK_TNFA_SIGNALING_VIA_NFKB   | 0.672| 1.234|   64|
#> |HALLMARK_ESTROGEN_RESPONSE_EARLY   | 0.779| 1.160|   72|
#> |HALLMARK_HEME_METABOLISM           | 0.779| 1.155|   84|
#> |HALLMARK_INTERFERON_GAMMA_RESPONSE | 0.808| 1.132|   61|
#> 
#> 
#> #### Downregulated
#> 
#> 
#> 
#> |pathway                            |  padj|    NES| size|
#> |:----------------------------------|-----:|------:|----:|
#> |HALLMARK_IL2_STAT5_SIGNALING       | 0.672| -1.377|   67|
#> |HALLMARK_KRAS_SIGNALING_UP         | 0.672| -1.288|   80|
#> |HALLMARK_OXIDATIVE_PHOSPHORYLATION | 0.672| -1.277|   77|
#> |HALLMARK_GLYCOLYSIS                | 0.672| -1.243|   72|
#> |HALLMARK_PEROXISOME                | 0.779| -1.201|   41|
#> |HALLMARK_TGF_BETA_SIGNALING        | 0.808| -1.172|   24|
#> 
#> 
#> ### treatment_D_vs_C
#> 
#> 
#> 
#> #### Upregulated
#> 
#> 
#> 
#> |pathway                                    |  padj|   NES| size|
#> |:------------------------------------------|-----:|-----:|----:|
#> |HALLMARK_PI3K_AKT_MTOR_SIGNALING           | 0.377| 1.575|   34|
#> |HALLMARK_INFLAMMATORY_RESPONSE             | 0.377| 1.436|   79|
#> |HALLMARK_XENOBIOTIC_METABOLISM             | 0.377| 1.423|   80|
#> |HALLMARK_UV_RESPONSE_DN                    | 0.477| 1.350|   58|
#> |HALLMARK_INTERFERON_GAMMA_RESPONSE         | 0.477| 1.331|   61|
#> |HALLMARK_ESTROGEN_RESPONSE_LATE            | 0.477| 1.311|   85|
#> |HALLMARK_EPITHELIAL_MESENCHYMAL_TRANSITION | 0.477| 1.306|   84|
#> |HALLMARK_ESTROGEN_RESPONSE_EARLY           | 0.526| 1.273|   72|
#> |HALLMARK_TNFA_SIGNALING_VIA_NFKB           | 0.526| 1.269|   64|
#> |HALLMARK_HEME_METABOLISM                   | 0.568| 1.206|   84|
#> 
#> 
#> #### Downregulated
#> 
#> 
#> 
#> |pathway                                  |  padj|    NES| size|
#> |:----------------------------------------|-----:|------:|----:|
#> |HALLMARK_PROTEIN_SECRETION               | 0.477| -1.375|   44|
#> |HALLMARK_BILE_ACID_METABOLISM            | 0.568| -1.232|   40|
#> |HALLMARK_ANDROGEN_RESPONSE               | 0.568| -1.229|   40|
#> |HALLMARK_OXIDATIVE_PHOSPHORYLATION       | 0.568| -1.210|   77|
#> |HALLMARK_TGF_BETA_SIGNALING              | 0.568| -1.186|   24|
#> |HALLMARK_MYOGENESIS                      | 0.568| -1.177|   72|
#> |HALLMARK_REACTIVE_OXYGEN_SPECIES_PATHWAY | 0.714| -1.072|   20|
#> |HALLMARK_HYPOXIA                         | 0.727| -1.038|   81|
#> |HALLMARK_MYC_TARGETS_V2                  | 0.788| -0.996|   20|
#> |HALLMARK_COAGULATION                     | 0.830| -0.969|   49|